HDDS-9062. NPE during investigation of datanode OOM.

+++ hadoop-hdds/container-service/src/main/java/org/apache/hadoop/ozone/container/replication/SendContainerRequestHandler.java
@@ -128,7 +128,9 @@ class SendContainerRequestHandler
 
   private void deleteTarball() {
     try {
-      Files.deleteIfExists(path);
+      if (null != path) {
+        Files.deleteIfExists(path);
+      }
     } catch (IOException e) {
       LOG.warn("Error removing {}", path);
     }
